+++ lib/src/io.dart
@@ -3,9 +3,11 @@ import 'dart:async';
 import 'package:file/file.dart';
 import 'package:glob/glob.dart';
 import 'package:glob/list_local_fs.dart';
-import 'package:path/path.dart' as p;
+import 'package:path/path.dart' as path;
 
 // TODO move into a util class https://github.com/jesims/docs/blob/master/Coding-Conventions/Dart.md#utilities
+import 'file_system_entity_extensions.dart';
+import 'file_system_entity_util.dart';
 import 'strings.dart';
 
 /// Lists all [File](https://pub.dev/packages/file) [FileSystemEntity]s matching the provided globs.
@@ -16,7 +18,7 @@ import 'strings.dart';
 ///
 /// [workingDirectory] will be the root path for the glob pattern to be
 /// evaluated in. Defaults to the current working directory provided by
-/// [p.context]
+/// [path.context]
 ///
 /// `since 0.0.1`
 //TODO add ability to specify FileSystem
@@ -25,8 +27,8 @@ Stream<FileSystemEntity> listFilesByGlob(
   String? workingDirectory,
 }) {
   var context = isNotBlank(workingDirectory)
-      ? p.Context(current: workingDirectory)
-      : p.context;
+      ? path.Context(current: workingDirectory)
+      : path.context;
   return Glob(
     globs.length == 1 ? globs.first : "{${globs.join(',')}}",
     context: context,
@@ -34,10 +36,9 @@ Stream<FileSystemEntity> listFilesByGlob(
 }
 
 /// Recursively deletes all [File](https://pub.dev/packages/file) [FileSystemEntity]s in the provided [Stream]
+/// using [FileSystemEntityUtil.delete]
 ///
 /// `since 0.0.1`
 Future<void> deleteFiles(Stream<FileSystemEntity> files) async {
-  return files.forEach((fse) async {
-    await fse.delete(recursive: true);
-  });
+  return files.forEach((fse) async => await fse.forceDelete(recursive: true));
 }