+++ test/require_file_test.dart
@@ -1,6 +1,6 @@
 import 'package:faker/faker.dart';
 import 'package:file/memory.dart';
-import 'package:pouch/pouch.dart' as p;
+import 'package:pouch/pouch.dart' as path;
 import 'package:pouch/pouch_io.dart';
 import 'package:test/test.dart';
 
@@ -12,7 +12,7 @@ void main() {
         expect(await file.exists(), false);
         expect(
           () => RequireFile.exists(file),
-          throwsA(TypeMatcher<p.RequireException>()),
+          throwsA(TypeMatcher<path.RequireException>()),
         );
       });
       test('returns the file if it does exist', () async {